Replace some functionality with the libgit api
