+++ SemDiff.Test/SemDiff.Test.csproj
@@ -144,7 +144,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PullRequestListTests.cs" />
     <Compile Include="Diff3Tests.cs" />
-    <Compile Include="GetRepoTests.cs" />
     <Compile Include="RepoTests.cs" />
     <Compile Include="FalsePositiveAnalysisTests.cs" />
     <Compile Include="TestBase.cs" />
@@ -174,15 +173,6 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Newtonsoft.Ancestor.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="testgitconfig3.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="testgitconfig2.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="testgitconfig.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
