--- SemDiff.Test/testgitconfig.txt
@@ -1,16 +1,0 @@
-[core]
-	bare = false
-	filemode = false
-	symlinks = false
-	ignorecase = true
-	logallrefupdates = true
-[core]
-	repositoryformatversion = 0
-[remote "origin"]
-	url = https://github.com/dotnet/roslyn.git
-[remote "origin"]
-	fetch = +refs/heads/*:refs/remotes/origin/*
-[branch "master"]
-	remote = origin
-[branch "master"]
-	merge = refs/heads/master
